fix(64bit): cast the remaining -Wsign-conversion sites (#11)
The non-rnd half of Phase 7 (31 sites), all representation-preserving
casts that match what the implicit conversion already did, so golden
output is byte-identical:

- qsort nmemb: the *_len() count (int) feeding qsort's size_t nmemb,
  cast (size_t) in bulk — use.c, buy.c, seed.c, input.c, gm.c (5),
  perm.c (4), report.c (3), swear.c, check.c.
- loop_known macro (olympia/loop.h): one (size_t) on its embedded
  qsort(ilist_len(kn)) clears all 6 expansion sites (summary.c,
  io.c, gm.c x2, report.c x2).
- mapgen/z.c: *((int *)p) = (int)size — the my_malloc size header
  slot (mapgen's my_malloc takes unsigned).
- perm.c / sout.c: the fixed `spaces` buffer — (size_t)spaces_len in
  the subscript and the my_malloc((size_t)spaces_len+1) sizing.

Golden gate YES (debug + asan-ubsan), asan/ubsan clean.

+++ mapgen/z.c
@@ -33,7 +33,7 @@ my_malloc(unsigned size)
 
 	bzero(p, size);
 
-	*((int *) p) = size;
+	*((int *) p) = (int)size;	/* size header slot, value fits */
 	*((int *) (p + size)) = 0xABCF;
 
 	return p + sizeof(int);
@@ -55,7 +55,7 @@ my_realloc(void *ptr, unsigned size)
 
 	p = realloc(p, size + sizeof(int));
 
-	*((int *)p) = size;
+	*((int *)p) = (int)size;	/* size header slot, value fits */
 	*((int *) (p + size)) = 0xABCF;
 
 	if (p == NULL)
