+++ olympia/input.c
@@ -443,7 +443,7 @@ sort_load_queue(ilist l)
 	for (i = 0; i < ilist_len(l); i++)
 		bx[l[i]]->temp = exec_precedence(l[i]);
 
-	qsort(l, ilist_len(l), sizeof(int), exec_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), exec_comp);
 }
 
 
@@ -470,7 +470,7 @@ sort_run_queue(ilist l)
 		bx[l[i]]->temp = pri * 1000000 + exec_precedence(l[i]);
 	}
 
-	qsort(l, ilist_len(l), sizeof(int), exec_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), exec_comp);
 }
